move setupOperatorContract() to test-utils
