+++ packages/sdk/src/exports.ts
@@ -93,9 +93,6 @@ import {
     delegate,
     undelegate,
     deploySponsorshipContract,
-    setupOperatorContract,
-    SetupOperatorContractOpts,
-    SetupOperatorContractReturnType,
     deployOperatorContract,
     DeployOperatorContractOpts,
     sponsor,
@@ -116,7 +113,6 @@ const _operatorContractUtils = {
     delegate,
     undelegate,
     deploySponsorshipContract,
-    setupOperatorContract,
     sponsor,
     stake,
     unstake,
@@ -127,7 +123,7 @@ const _operatorContractUtils = {
     getOperatorContract
 }
 export { _operatorContractUtils }
-export type { SetupOperatorContractOpts, SetupOperatorContractReturnType, DeployOperatorContractOpts, DeploySponsorshipContractOpts }
+export type { DeployOperatorContractOpts, DeploySponsorshipContractOpts }
 
 export type { IceServer, PeerDescriptor, PortRange } from '@streamr/dht'
 export type { AbstractSigner, Eip1193Provider, Overrides } from 'ethers'
